+++ Release notes.md
@@ -11,7 +11,7 @@
     - `recognizerCollection`, which is a collection of Recognizers used for card scanning.
     - `frontImage`, which would represent the image of the card where the card number in located in the Base64 format string
     - `backImage`,  which would represent the image of the second side of the card in the Base64 format string
-        - the `backImage` parameter is optional when scanning the card that contains all of the information on one side (or if you extract specific information located only on one side), and can be passed as `null` or an empty string (`””`). 
+        - the `backImage` parameter is optional when scanning the card that contains all of the information on one side (or if you extract specific information located only on one side), and can be left out from the implementation or passed as an empty string (`””`) 
 - An example of its usage can be found in the [sample application](https://github.com/blinkcard/blinkcard-capacitor/blob/main/sample_files/home.page.ts) , both for the two-sided and one-sided card scanning. 
 - More information about the DirectAPI scanning can be found here in the native documentation for [Android](https://github.com/BlinkCard/blinkcard-android?tab=readme-ov-file#direct-api) and [iOS](https://github.com/BlinkCard/blinkcard-ios?tab=readme-ov-file#direct-api-processing)
 - We still recommend using our ‘regular’ way of scanning with the camera, as static images can sometimes be in lower-quality which can cause SDK extraction error. It would be best to use the `scanWithDirectApi` method when using the device’s camera is not an option.