+++ Release notes.md
@@ -1,6 +1,7 @@
 ## 2.9.1
 
 - Updated to [Android SDK v2.9.3](https://github.com/blinkcard/blinkcard-android/releases/tag/v2.9.3) and [iOS SDK v2.9.1](https://github.com/BlinkCard/blinkcard-ios/releases/tag/v2.9.1)
+- This version of the SDK contains the native iOS BlinkCard.xcframework with the privacy manifest file (`PrivacyInfo.xcprivacy`).
 
 ### Major API update
 
@@ -14,7 +15,7 @@
         - the `backImage` parameter is optional when scanning the card that contains all of the information on one side (or if you extract specific information located only on one side), and can be left out from the implementation or passed as an empty string (`””`) 
 - An example of its usage can be found in the [sample application](https://github.com/blinkcard/blinkcard-capacitor/blob/main/sample_files/home.page.ts) , both for the two-sided and one-sided card scanning. 
 - More information about the DirectAPI scanning can be found here in the native documentation for [Android](https://github.com/BlinkCard/blinkcard-android?tab=readme-ov-file#direct-api) and [iOS](https://github.com/BlinkCard/blinkcard-ios?tab=readme-ov-file#direct-api-processing)
-- We still recommend using our ‘regular’ way of scanning with the camera, as static images can sometimes be in lower-quality which can cause SDK extraction error. It would be best to use the `scanWithDirectApi` method when using the device’s camera is not an option.
+- We still recommend using direct camera scanning, as static images can sometimes be in lower-quality which can cause SDK extraction error. It would be best to use the scanWithDirectApi method when using the device’s camera is not an option.
 
 ## 2.9.0